src/web/requirements.txt
coverage==5.0.1
Django==3.0.2
django-environ==0.4.5
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.1
sqlparse==0.3.0
uWSGI==2.0.18
djangorestframework==3.11.0
